Moved deviceTrustPage to security
