+++ src/app/app-routing.module.ts
@@ -36,7 +36,6 @@ const routes: Routes = [
   { path: 'profile', loadChildren: './pages/profile/profile.module#ProfilePageModule' },
   { path: 'security', loadChildren: './pages/security/security.module#SecurityPageModule' },
   { path: 'files', loadChildren: './pages/files/files.module#FilesPageModule' },
-  { path: 'security/deviceTrust', loadChildren: './pages/security/deviceTrust/deviceTrust.module#DeviceTrustPageModule' },
 ];
 @NgModule({
   imports: [
