+++ src/app/app.component.html
@@ -53,30 +53,6 @@
                 Device Security
               </ion-label>
             </ion-item>
-            <ion-item routerDirection='root' routerLink="/security">
-              <ion-icon slot="start"></ion-icon>
-              <ion-label>
-                Documentation
-              </ion-label>
-            </ion-item>
-            <ion-item routerDirection='root' routerLink="/security/deviceTrust">
-              <ion-icon slot="start"></ion-icon>
-              <ion-label>
-                Device Trust
-              </ion-label>
-            </ion-item>
-            <ion-item routerDirection='root' routerLink="/security">
-              <ion-icon slot="start"></ion-icon>
-              <ion-label>
-                Secure Storage
-              </ion-label>
-            </ion-item>
-            <ion-item routerDirection='root' routerLink="/security">
-              <ion-icon slot="start"></ion-icon>
-              <ion-label>
-                Cert Pinning
-              </ion-label>
-            </ion-item>
           </ion-menu-toggle>
           <ion-menu-toggle auto-hide="false">
             <ion-item routerDirection='root' routerLink="/docs">
